- Refactoring: removing ResultsValidator.java, moving ParsedJavaFile.java to model package
